Fixed race condition with calculation for left and right coordinates of map view to calculate length
This caused the cell size sometimes to be 0 due to a race condition when converting coordinates from the map view to its super view. When that happened, all annotations become invisible. This constantly happens when rotating, especially in flyover mode.
Instead we are now calculating it against its own view, which always returns consistent and valid results.