Fix NSInternalInconsistencyException related to KVO
When the map view is deallocated before the delegate proxy, the proxy can't remove itself from the map view's KVO observers and an `NSInternalInconsistencyException` was thrown ("An instance … of class MKMapView was deallocated while key value observers were still registered with it. Current observation"). The issue was reproducible on macOS 10.12.6/Xcode 9.2 by running the unit test `testDeallocMapView`.

To fix this problem, the proxy now references the map view strongly so that the removal can happen in its `dealloc` method. Fixes #136